refactor(guild/group): extract role list related logic
