+++ src/components/[guild]/Roles.tsx
@@ -1,0 +1,95 @@
+import { Center, Spinner, Stack } from "@chakra-ui/react"
+import CollapsibleRoleSection from "components/[guild]/CollapsibleRoleSection"
+import useGuild from "components/[guild]/hooks/useGuild"
+import { RequirementErrorConfigProvider } from "components/[guild]/Requirements/RequirementErrorConfigContext"
+import RoleCard from "components/[guild]/RoleCard/RoleCard"
+import useScrollEffect from "hooks/useScrollEffect"
+import dynamic from "next/dynamic"
+import { useMemo, useRef, useState } from "react"
+import { Visibility } from "types"
+
+const BATCH_SIZE = 10
+
+const DynamicNoRolesAlert = dynamic(() => import("components/[guild]/NoRolesAlert"))
+
+const Roles = () => {
+  const { roles: allRoles } = useGuild()
+  const roles = allRoles.filter((role) => !role.groupId)
+
+  // temporary, will order roles already in the SQL query in the future
+  const sortedRoles = useMemo(() => {
+    if (roles?.every((role) => role.position === null)) {
+      const byMembers = roles?.sort(
+        (role1, role2) => role2.memberCount - role1.memberCount
+      )
+      return byMembers
+    }
+
+    return (
+      roles?.sort((role1, role2) => {
+        if (role1.position === null) return 1
+        if (role2.position === null) return -1
+        return role1.position - role2.position
+      }) ?? []
+    )
+  }, [roles])
+
+  const publicRoles = sortedRoles.filter(
+    (role) => role.visibility !== Visibility.HIDDEN
+  )
+  const hiddenRoles = sortedRoles.filter(
+    (role) => role.visibility === Visibility.HIDDEN
+  )
+
+  // TODO: we use this behaviour in multiple places now, should make a useScrollBatchedRendering hook
+  const [renderedRolesCount, setRenderedRolesCount] = useState(BATCH_SIZE)
+  const rolesEl = useRef(null)
+  useScrollEffect(() => {
+    if (
+      !rolesEl.current ||
+      rolesEl.current.getBoundingClientRect().bottom > window.innerHeight ||
+      roles?.length <= renderedRolesCount
+    )
+      return
+
+    setRenderedRolesCount((prevValue) => prevValue + BATCH_SIZE)
+  }, [roles, renderedRolesCount])
+
+  return (
+    <>
+      {publicRoles.length ? (
+        <RequirementErrorConfigProvider>
+          <Stack ref={rolesEl} spacing={4}>
+            {publicRoles.map((role, i) => {
+              if (i > renderedRolesCount - 1) return null
+              return <RoleCard key={role.id} role={role} />
+            })}
+          </Stack>
+        </RequirementErrorConfigProvider>
+      ) : (
+        <DynamicNoRolesAlert />
+      )}
+
+      {publicRoles?.length && roles?.length > renderedRolesCount && (
+        <Center pt={6}>
+          <Spinner />
+        </Center>
+      )}
+
+      {!!hiddenRoles?.length && (
+        <CollapsibleRoleSection
+          id="hiddenRoles"
+          roleCount={hiddenRoles.length}
+          label="hidden"
+          defaultIsOpen
+        >
+          {hiddenRoles.map((role) => (
+            <RoleCard key={role.id} role={role} />
+          ))}
+        </CollapsibleRoleSection>
+      )}
+    </>
+  )
+}
+
+export default Roles
